requirements.txt
flask
python-dotenv
twilio
google-generativeai
requests
gunicorn
numpy>=1.21.0
scikit-learn>=1.0.0
# OCR Dependencies
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0